Extract HTTP message parsing
